feat : add first contract and second contract and deployed as proxy
