+++ package.json
@@ -1,23 +1,13 @@
 {
-  "name": "assignment",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/shivam7868-alt/Blockchain-assignment.git"
-  },
-  "author": "",
-  "license": "ISC",
-  "bugs": {
-    "url": "https://github.com/shivam7868-alt/Blockchain-assignment/issues"
-  },
-  "homepage": "https://github.com/shivam7868-alt/Blockchain-assignment#readme",
+  "name": "hardhat-project",
   "devDependencies": {
     "@nomicfoundation/hardhat-toolbox": "^2.0.2",
-    "hardhat": "^2.13.0"
+    "@nomiclabs/hardhat-ethers": "^2.2.3",
+    "@openzeppelin/hardhat-upgrades": "^1.22.1",
+    "ethers": "^5.7.2",
+    "hardhat": "^2.13.1"
+  },
+  "dependencies": {
+    "dotenv": "^16.0.3"
   }
 }
