--- scripts/deploy.js
@@ -1,23 +1,0 @@
-// We require the Hardhat Runtime Environment explicitly here. This is optional
-// but useful for running the script in a standalone fashion through `node <script>`.
-//
-// You can also run a script with `npx hardhat run <script>`. If you do that, Hardhat
-// will compile your contracts, add the Hardhat Runtime Environment's members to the
-// global scope, and execute the script.
-const hre = require("hardhat");
-
-async function main() {
-  const bContract = await ethers.getContractFactory("B");
-        const B = await bContract.deploy();
-        await B.deployed();
-        const aContract = await ethers.getContractFactory("A");
-        const A = await aContract.deploy(B.address);
-        await A.deployed();
-}
-
-// We recommend this pattern to be able to use async/await everywhere
-// and properly handle errors.
-main().catch((error) => {
-  console.error(error);
-  process.exitCode = 1;
-});
